\name{ExportSoilpropsGrids}
\alias{ExportSoilpropsGrids}
\title{Export grids of soil properties.}

\description{Export all gripds of soil properties to ./euptf_attributes and ./euptf_rasters}

\usage{ExportSoilpropsGrids()
}
\arguments{
  \item{}{
	}
  }
\value{None.
}

\details{Export all gripds of soil properties to ./euptf_attributes and ./euptf_rasters. Only optional for further workflow in lumpR.

}
\references{
}
\author{Sophia Dobkowitz <dobkowit@uni-potsdam.de>}

\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
ExportSoilpropsGrids()
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~soilgrids }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~properties }% __ONLY ONE__ keyword per line
\keyword{ ~raster }% __ONLY ONE__ keyword per line
